move aws promts to new file
